+++ coinlib/CHANGELOG.md
@@ -5,7 +5,11 @@ secp256k1-coinlib fork.
 
 - Adds MuSig2 support with `MuSigPublicKeys` and `MuSigStatefulSigningSession`.
 - Includes adaptor signature support when using MuSig2 with
-- `SchnorrAdaptorSignature`.
+  `SchnorrAdaptorSignature`.
+- Adds `InputSequence` abstraction for input sequences. Input sequences are now
+    `0xfffffffe` by default and enforce locktimes.
+- Adds `locktimeIsEnforced` to `Transaction` to determine if the locktime is in
+    effect.
 - Moves to underlying secp256k1-coinlib.
 - Removed dependency to wasm_interop that had a broken js dependency.
 - Fixes `extraEntropy` being ignored for `Secp256k1Base.schnorrSign`.